test: use test runner
