+++ examples/complete/fixtures.us-east-2.tfvars
@@ -1,5 +1,3 @@
-enabled = true
-
 region = "us-east-2"
 
 availability_zones = ["us-east-2a", "us-east-2b"]
