+++ test/src/examples_complete_test.go
@@ -1,59 +1,15 @@
 package test
 
 import (
-	"os"
-	"strings"
+	"regexp"
 	"testing"
 
-	"github.com/gruntwork-io/terratest/modules/logger"
-	"github.com/gruntwork-io/terratest/modules/random"
 	"github.com/gruntwork-io/terratest/modules/terraform"
-	testStructure "github.com/gruntwork-io/terratest/modules/test-structure"
 	"github.com/stretchr/testify/assert"
-	"k8s.io/apimachinery/pkg/util/runtime"
 )
 
 // Test the Terraform module in examples/complete using Terratest.
-func TestExamplesComplete(t *testing.T) {
-	t.Parallel()
-
-	randID := strings.ToLower(random.UniqueId())
-	attributes := []string{randID}
-
-	rootFolder := "../../"
-	terraformFolderRelativeToRoot := "examples/complete"
-	varFiles := []string{"fixtures.us-east-2.tfvars"}
-
-	tempTestFolder := testStructure.CopyTerraformFolderToTemp(t, rootFolder, terraformFolderRelativeToRoot)
-
-	// We need to create the ALB first because terraform does not wwait for it to be in the ready state before creating ECS target group
-	terraformOptions := &terraform.Options{
-		// The path to where our Terraform code is located
-		TerraformDir: tempTestFolder,
-		Upgrade:      true,
-		// Variables to pass to our Terraform code using -var-file options
-		VarFiles: varFiles,
-		Vars: map[string]interface{}{
-			"attributes": attributes,
-		},
-	}
-
-	// Keep the output quiet
-	if !testing.Verbose() {
-		terraformOptions.Logger = logger.Discard
-	}
-
-	// At the end of the test, run `terraform destroy` to clean up any resources that were created
-	defer cleanup(t, terraformOptions, tempTestFolder)
-
-	// If Go runtime crashes, run `terraform destroy` to clean up any resources that were created
-	defer runtime.HandleCrash(func(i interface{}) {
-		terraform.Destroy(t, terraformOptions)
-	})
-
-	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
-	terraform.InitAndApply(t, terraformOptions)
-
+func testExamplesComplete(t *testing.T, terraformOptions *terraform.Options, randID string, _ string) {
 	// Run `terraform output` to get the value of an output variable
 	vpcCidr := terraform.Output(t, terraformOptions, "vpc_cidr")
 	// Verify we're getting back the outputs we expect
@@ -90,8 +46,10 @@ func TestExamplesComplete(t *testing.T) {
 	assert.Contains(t, readerEndpoint, "eg-test-documentdb-cluster-"+randID+".cluster-ro")
 }
 
-func cleanup(t *testing.T, terraformOptions *terraform.Options, tempTestFolder string) {
-	terraform.Destroy(t, terraformOptions)
-	err := os.RemoveAll(tempTestFolder)
-	assert.NoError(t, err)
+func testExamplesCompleteDisabled(t *testing.T, terraformOptions *terraform.Options, randID string, results string) {
+	// Should complete successfully without creating or changing any resources.
+	// Extract the "Resources:" section of the output to make the error message more readable.
+	re := regexp.MustCompile(`Resources: [^.]+\.`)
+	match := re.FindString(results)
+	assert.Equal(t, "Resources: 0 added, 0 changed, 0 destroyed.", match, "Re-applying the same configuration should not change any resources")
 }
